PanelDue, Thumbnail: moved thumbnail structs into seperate module
